requirements.txt
tensorflow
pandas
numpy
scipy
opencv-python
ipykernel
jupyter
jupyterlab
matplotlib
seaborn
Flask
Pillow